Improve and test error message
Hand the bulk of error message generation to functions in the brackets
namespace with the added benefit of making that code testable. Implement
those tests.

Some of the messages are still a little rough around the edges and the
unneeded punctuation at the end of the stream error and all's well
messages is annoying, but the code seems to work.

Meets spec 3.